Moving spacing file.
