+++ src/pages/buttonGroup.tsx
@@ -6,7 +6,7 @@ import Grid from "../components/Grid/Grid";
 import PageTitle from "../components_site/PageTitle/PageTitle";
 import { classes as text } from "../styles/default/text.st.css";
 // import { classes as grid } from "../styles/default/grid.st.css";
-import { classes as spacing } from "../styles/shelley/mixins/spacing.st.css";
+import { classes as spacing } from "../styles/default/spacing.st.css";
 import Blockquote from "../components/Blockquote/Blockquote";
 import {
   meta,
